@@ -3,7 +3,7 @@ requires = [
     "hatchling",
     "hatch-vcs",
     "param >=2.1.0",
-    "bokeh >=3.4.0,<3.5.0",
+    "bokeh ==3.5.0.dev7",
     "pyviz_comms >=0.7.4",
     "requests",
     "packaging",
@@ -16,7 +16,7 @@ dynamic = ["version"]
 description = 'The powerful data exploration & web app framework for Python.'
 readme = "README.md"
 license = { text = "BSD" }
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 authors = [{ name = "HoloViz developers", email = "developers@holoviz.org" }]
 maintainers = [
     { name = "HoloViz developers", email = "developers@holoviz.org" },
@@ -48,7 +48,7 @@ classifiers = [
 ]
 
 dependencies = [
-    'bokeh >=3.4.0,<3.5.0',
+    'bokeh ==3.5.0.dev7',
     'param >=2.1.0,<3.0',
     'pyviz_comms >=2.0.0',
     'markdown',
@@ -144,6 +144,7 @@ ignore = [
     "W605", # Invalid escape sequence
     "E701", # Multiple statements on one line
     "B006", # Do not use mutable data structures for argument defaults
+    "B905", # `zip()` without an explicit `strict=` parameter
 ]
 select = [
     "B",
