@@ -43,6 +43,11 @@ html_css_files += [
 html_theme = "pydata_sphinx_theme"
 html_favicon = "_static/icons/favicon.ico"
 
+current_release = panel.__version__  # Current release version variable
+
+announcement_text = f"Panel {current_release} has just been released! Check out the <a href='https://panel.holoviz.org/about/releases.html#version-1-4-0'>release notes</a> and support Panel by giving it a 🌟 on <a href='https://github.com/holoviz/panel'>Github</a>."
+
+
 html_theme_options = {
     "logo": {
         "image_light": "_static/logo_horizontal_light_theme.png",
@@ -74,7 +79,7 @@ html_theme_options = {
         "panelitelink",
         "page-toc",
     ],
-    "announcement": "Panel 1.4 has just been released! Checkout the <a href='https://panel.holoviz.org/about/releases.html#version-1-4-0'>release notes</a> and support Panel by giving it a 🌟 on <a href='https://github.com/holoviz/panel'>Github</a>.",
+    "announcement": announcement_text,
 }
 
 extensions += [
