@@ -27,11 +27,11 @@ export class TrendIndicatorView extends HTMLBoxView {
 
   override initialize(): void {
     super.initialize()
-    this.containerDiv = div({style: "height:100%; width:100%;"})
-    this.titleDiv = div({style: "font-size: 1em; word-wrap: break-word;"})
-    this.valueDiv = div({style: "font-size: 2em"})
-    this.value2Div = div({style: "font-size: 1em; opacity: 0.5; display: inline"})
-    this.changeDiv = div({style: "font-size: 1em; opacity: 0.5; display: inline"})
+    this.containerDiv = div({style: {height: "100%", width: "100%"}})
+    this.titleDiv = div({style: {font_size: "1em", word_wrap: "break-word"}})
+    this.valueDiv = div({style: {font_size: "2em"}})
+    this.value2Div = div({style: {font_size: "1em", opacity: "0.5", display: "inline"}})
+    this.changeDiv = div({style: {font_size: "1em", opacity: "0.5", display: "inline"}})
     this.textDiv = div({}, this.titleDiv, this.valueDiv, div({}, this.changeDiv, this.value2Div))
 
     this.updateTitle()
@@ -41,7 +41,7 @@ export class TrendIndicatorView extends HTMLBoxView {
     this.updateTextFontSize()
 
     this.plotDiv = div({})
-    this.containerDiv = div({style: "height:100%; width:100%"}, this.textDiv, this.plotDiv)
+    this.containerDiv = div({style: {height: "100%", width: "100%"}}, this.textDiv, this.plotDiv)
     this.updateLayout()
   }
 
