@@ -24,7 +24,9 @@ from panel.pane import (
 from panel.param import (
     JSONInit, Param, ParamFunction, ParamMethod, Skip,
 )
-from panel.tests.util import mpl_available, mpl_figure
+from panel.tests.util import (
+    async_wait_until, mpl_available, mpl_figure, wait_until,
+)
 from panel.widgets import (
     AutocompleteInput, Button, Checkbox, DatePicker, DatetimeInput,
     EditableFloatSlider, EditableRangeSlider, LiteralInput, NumberInput,
@@ -1789,11 +1791,11 @@ def test_param_generator(document, comm):
 
     root = pane.get_root(document, comm)
 
-    assert root.children[0].text == '&lt;p&gt;False&lt;/p&gt;\n'
+    wait_until(lambda: root.children[0].text == '&lt;p&gt;False&lt;/p&gt;\n')
 
     checkbox.value = True
 
-    assert root.children[0].text == '&lt;p&gt;True&lt;/p&gt;\n'
+    wait_until(lambda: root.children[0].text == '&lt;p&gt;True&lt;/p&gt;\n')
 
 
 def test_param_generator_append(document, comm):
@@ -1807,15 +1809,17 @@ def test_param_generator_append(document, comm):
 
     root = pane.get_root(document, comm)
 
-    assert len(root.children) == 2
+    wait_until(lambda: len(root.children) == 2)
     assert root.children[0].text == '&lt;p&gt;False&lt;/p&gt;\n'
     assert root.children[1].text == '&lt;p&gt;True&lt;/p&gt;\n'
 
     checkbox.value = True
 
-    assert len(root.children) == 2
-    assert root.children[0].text == '&lt;p&gt;True&lt;/p&gt;\n'
-    assert root.children[1].text == '&lt;p&gt;False&lt;/p&gt;\n'
+    wait_until(lambda: len(root.children) == 2)
+    wait_until(lambda: (
+        (root.children[0].text == '&lt;p&gt;True&lt;/p&gt;\n') and
+        (root.children[1].text == '&lt;p&gt;False&lt;/p&gt;\n')
+    ))
 
 async def test_param_async_generator(document, comm):
     checkbox = Checkbox(value=False)
@@ -1827,15 +1831,11 @@ async def test_param_async_generator(document, comm):
 
     root = pane.get_root(document, comm)
 
-    await asyncio.sleep(0.01)
-
-    assert root.children[0].text == '&lt;p&gt;False&lt;/p&gt;\n'
+    await async_wait_until(lambda: root.children[0].text == '&lt;p&gt;False&lt;/p&gt;\n')
 
     checkbox.value = True
 
-    await asyncio.sleep(0.01)
-
-    assert root.children[0].text == '&lt;p&gt;True&lt;/p&gt;\n'
+    await async_wait_until(lambda: root.children[0].text == '&lt;p&gt;True&lt;/p&gt;\n')
 
 async def test_param_async_generator_append(document, comm):
     checkbox = Checkbox(value=False)
@@ -1849,21 +1849,17 @@ async def test_param_async_generator_append(document, comm):
 
     root = pane.get_root(document, comm)
 
-    await asyncio.sleep(0.01)
-    assert len(root.children) == 1
-    assert root.children[0].text == '&lt;p&gt;False&lt;/p&gt;\n'
-    await asyncio.sleep(0.01)
-    assert len(root.children) == 2
+    await async_wait_until(lambda: len(root.children) == 1, interval=10)
+    await async_wait_until(lambda: root.children[0].text == '&lt;p&gt;False&lt;/p&gt;\n')
+    await async_wait_until(lambda: len(root.children) == 2, interval=10)
     assert root.children[0].text == '&lt;p&gt;False&lt;/p&gt;\n'
     assert root.children[1].text == '&lt;p&gt;True&lt;/p&gt;\n'
 
     checkbox.value = True
 
-    await asyncio.sleep(0.01)
-    assert len(root.children) == 1
+    await async_wait_until(lambda: len(root.children) == 1, interval=10)
     assert root.children[0].text == '&lt;p&gt;True&lt;/p&gt;\n'
-    await asyncio.sleep(0.01)
-    assert len(root.children) == 2
+    await async_wait_until(lambda: len(root.children) == 2, interval=10)
     assert root.children[0].text == '&lt;p&gt;True&lt;/p&gt;\n'
     assert root.children[1].text == '&lt;p&gt;False&lt;/p&gt;\n'
 
@@ -1879,11 +1875,11 @@ def test_param_generator_multiple(document, comm):
 
     root = pane.get_root(document, comm)
 
-    assert root.children[0].text == '&lt;p&gt;True&lt;/p&gt;\n'
+    wait_until(lambda: root.children[0].text == '&lt;p&gt;True&lt;/p&gt;\n')
 
     checkbox.value = True
 
-    assert root.children[0].text == '&lt;p&gt;False&lt;/p&gt;\n'
+    wait_until(lambda: root.children[0].text == '&lt;p&gt;False&lt;/p&gt;\n')
 
 async def test_param_async_generator_multiple(document, comm):
     checkbox = Checkbox(value=False)
