@@ -1,6 +1,8 @@
 import asyncio
 import datetime as dt
 
+from zoneinfo import ZoneInfo
+
 import numpy as np
 import pandas as pd
 import pytest
@@ -17,7 +19,6 @@ from panel.depends import bind
 from panel.io.state import set_curdoc
 from panel.models.tabulator import CellClickEvent, TableEditEvent
 from panel.tests.util import mpl_available, serve_and_request, wait_until
-from panel.util import BOKEH_JS_NAT
 from panel.widgets import Button, TextInput
 from panel.widgets.tables import DataFrame, Tabulator
 
@@ -336,7 +337,7 @@ def test_tabulator_expanded_content_pagination(document, comm):
     assert len(model.children) == 0
 
 
-def test_tabulator_expanded_content_embed(document, comm):
+def test_tabulator_content_embed(document, comm):
     df = makeMixedDataFrame()
 
     table = Tabulator(df, embed_content=True, row_content=lambda r: r.A)
@@ -356,6 +357,34 @@ def test_tabulator_expanded_content_embed(document, comm):
         assert i in model.children
         row = model.children[i]
         assert row.text  == f"&lt;pre&gt;{r.A+1}&lt;/pre&gt;"
+
+
+def test_tabulator_content_embed_and_expand(document, comm):
+    # https://github.com/holoviz/panel/issues/6200
+    df = makeMixedDataFrame()
+
+    calls = []
+    def row_content(row):
+        calls.append(row)
+        return row.A
+
+    table = Tabulator(df, embed_content=True, row_content=row_content)
+
+    model = table.get_root(document, comm)
+
+    assert len(calls) == len(df)
+
+    assert len(model.children) == len(df)
+
+    for i, r in df.iterrows():
+        assert i in model.children
+        row = model.children[i]
+        assert row.text  == f"&lt;pre&gt;{r.A}&lt;/pre&gt;"
+
+    # Expanding a row should not call row_content again in this context.
+    table.expanded = [1]
+
+    assert len(calls) == len(df)
 
 
 def test_tabulator_selected_and_filtered_dataframe(document, comm):
@@ -1327,16 +1356,16 @@ def test_tabulator_patch_with_NaT(document, comm):
     table.patch({'A': [(0, pd.NaT)]})
 
     # We're also checking that the NaT value that was in the original table
-    # at .loc[1, 'A'] is converted in the model as BOKEH_JS_NAT.
+    # at .loc[1, 'A'] is converted in the model as np.nan.
     expected = {
         'index': np.array([0, 1]),
-        'A': np.array([BOKEH_JS_NAT, BOKEH_JS_NAT])
+        'A': np.array([np.nan, np.nan])
     }
     for col, values in model.source.data.items():
         expected_array = expected[col]
         np.testing.assert_array_equal(values, expected_array)
         # Not checking that the data in table.value is the same as expected
-        # In table.value we have NaT values, in expected the BOKEH_JS_NAT constant.
+        # In table.value we have NaT values, in expected np.nan.
 
 
 def test_tabulator_stream_series_paginated_not_follow(document, comm):
@@ -1972,6 +2001,40 @@ def test_server_edit_event():
     wait_until(lambda: len(events) == 1)
     assert events[0].value == 3.14
     assert events[0].old == 1
+
+
+def test_edit_with_datetime_aware_column():
+    # https://github.com/holoviz/panel/issues/6673
+
+    # The order of these columns matter, 'B' and 'C' should be first as it's in fact
+    # processed first when 'A' is edited.
+    data = {
+        "B": pd.date_range(start='2024-01-01', end='2024-01-03', freq='D', tz='utc'),
+        "C": pd.date_range(start='2024-01-01', end='2024-01-03', freq='D', tz=ZoneInfo('US/Eastern')),
+        "A": ['a', 'b', 'c'],
+    }
+    df = pd.DataFrame(data)
+
+    table = Tabulator(df)
+
+    serve_and_request(table)
+
+    wait_until(lambda: bool(table._models))
+    ref, (model, _) = list(table._models.items())[0]
+    doc = list(table._documents.keys())[0]
+
+    events = []
+    table.on_edit(lambda e: events.append(e))
+
+    new_data = dict(model.source.data)
+    new_data['A'][1] = 'new'
+
+    table._server_change(doc, ref, None, 'data', model.source.data, new_data)
+    table._server_event(doc, TableEditEvent(model, 'A', 1))
+
+    wait_until(lambda: len(events) == 1)
+    assert events[0].value == 'new'
+    assert events[0].old == 'b'
 
 def test_tabulator_cell_click_event():
     df = makeMixedDataFrame()
